Add PGP configuration
settings.xml on build server has been updated to declare a 'pgp' profile
and keyname and passphrase for buildmaster@springframework.org. This
profile is active by default on the build server at all times.

This change creates PGP signatures using the maven-gpg-plugin if the
'pgp' profile is active.  Users can run it locally, but will have
to provide information about keyname and passphrase.